accepts decimal numbers for calculation

+++ assignments/week08/day-3/routes/calculate.js
@@ -4,8 +4,8 @@ const Model = require('./../models/Calculate.js');
 //methods of calc
 //addition
 CalcController.add = function(req, res) {
-    var num1 = parseInt(req.body.num1);
-    var num2 = parseInt(req.body.num2);
+    var num1 = parseFloat(req.body.num1);
+    var num2 = parseFloat(req.body.num2);
 
     Model.add(num1, num2, function(error, result) {
        if(error) {
@@ -18,8 +18,8 @@ CalcController.add = function(req, res) {
 
 //subtraction
 CalcController.sub = function(req, res) {
-    var num1 = parseInt(req.body.num1);
-    var num2 = parseInt(req.body.num2);
+    var num1 = parseFloat(req.body.num1);
+    var num2 = parseFloat(req.body.num2);
 
     Model.sub(num1, num2, function(error, result) {
        if(error) {
@@ -32,8 +32,8 @@ CalcController.sub = function(req, res) {
 
 //multiplication
 CalcController.mul = function(req, res) {
-    var num1 = parseInt(req.body.num1);
-    var num2 = parseInt(req.body.num2);
+    var num1 = parseFloat(req.body.num1);
+    var num2 = parseFloat(req.body.num2);
 
     Model.mul(num1, num2, function(error, result) {
        if(error) {
@@ -46,8 +46,8 @@ CalcController.mul = function(req, res) {
 
 //division
 CalcController.div = function(req, res) {
-    var num1 = parseInt(req.body.num1);
-    var num2 = parseInt(req.body.num2);
+    var num1 = parseFloat(req.body.num1);
+    var num2 = parseFloat(req.body.num2);
 
     Model.div(num1, num2, function(error, result) {
        if(error) {
